pyproject.toml
[tool.poetry]
name = "awcy-nfo"
version = "0.1.0"
description = "fresh baked readme's from your friends at awcy?"
authors = ["recce <small_teams_behind_enemy_lines@protonmail.com>"]
license = "BSD"
readme = "README.md"
include = ["headers", "styles", "docs"]

[tool.poetry.dependencies]
python = "^3.7.8"
click = "^7.1.2"
importlib_metadata = "^3.10.1"
"ruamel.yaml" = "^0.17.4"
colorama = "^0.4.4"
click-logging = "^1.0.1"
pandas = "^1.2.4"
pyinstaller = "^4.5.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^3.9.0"
black = "^20.8b1"
ipython = "^7.22.0"

[tool.poetry.scripts]
awcynfo = "awcy_nfo.cli:create_readme"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
